feat(remove-board-view): complete removal of board view, converted to use BoardJSON instead
