+++ lib/epochtalk_server_web/endpoint.ex
@@ -22,7 +22,7 @@ defmodule EpochtalkServerWeb.Endpoint do
     at: "/",
     from: :epochtalk_server,
     gzip: false,
-    only: ~w(assets fonts images favicon.ico robots.txt)
+    only: EpochtalkServerWeb.static_paths()
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.
