--- test/epochtalk_server_web/views/board_view_test.exs
@@ -1,7 +1,0 @@
-defmodule EpochtalkServerWeb.BoardViewTest do
-  use EpochtalkServerWeb.ConnCase, async: true
-  alias EpochtalkServerWeb.BoardView
-
-  # Specify that we want to use doctests:
-  doctest BoardView
-end
